Update dkms.conf to only build the package for kernel versions 6.1.x - 6.5.x
Proxmox 8.2.2 comes with kernel 6.8, which is unfortunately incompatible with this dkms module. To be able to install proxmox and its dependencys, this dkms module should skip the build if the kernel version is not adequate. This is exactly what this commit does - making sure this dkms module is only build for compatible kernel versions.